basic primitives and images

+++ netcanv-renderer/src/lib.rs
@@ -38,15 +38,15 @@ pub trait Image {
    fn colorized(&self, color: Color) -> Self;
 
    /// Returns the size of the image.
-   fn size(&self) -> (usize, usize);
+   fn size(&self) -> (u32, u32);
 
    /// Returns the width of the image.
-   fn width(&self) -> usize {
+   fn width(&self) -> u32 {
       self.size().0
    }
 
    /// Returns the height of the image.
-   fn height(&self) -> usize {
+   fn height(&self) -> u32 {
       self.size().1
    }
 }
